requirements.txt
click==6.6
Flask==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycrypto==2.6.1
Werkzeug==0.11.11
